chore: remove old cache command
